pyproject.toml
[project]
name = "rasan-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "none"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "markdown (>=3.10,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
